+++ common/rdm/VariableFieldSizeCalculatorTest.cpp
@@ -156,28 +156,28 @@ void VariableFieldSizeCalculatorTest::testStringFields() {
         3,
         &descriptor,
         &variable_field_size));
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(0), variable_field_size);
+  CPPUNIT_ASSERT_EQUAL(0u, variable_field_size);
   CPPUNIT_ASSERT_EQUAL(
       VariableFieldSizeCalculator::VARIABLE_STRING,
       m_calculator.CalculateFieldSize(
         4,
         &descriptor,
         &variable_field_size));
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(1), variable_field_size);
+  CPPUNIT_ASSERT_EQUAL(1u, variable_field_size);
   CPPUNIT_ASSERT_EQUAL(
       VariableFieldSizeCalculator::VARIABLE_STRING,
       m_calculator.CalculateFieldSize(
         34,
         &descriptor,
         &variable_field_size));
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(31), variable_field_size);
+  CPPUNIT_ASSERT_EQUAL(31u, variable_field_size);
   CPPUNIT_ASSERT_EQUAL(
       VariableFieldSizeCalculator::VARIABLE_STRING,
       m_calculator.CalculateFieldSize(
         35,
         &descriptor,
         &variable_field_size));
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(32), variable_field_size);
+  CPPUNIT_ASSERT_EQUAL(32u, variable_field_size);
   CPPUNIT_ASSERT_EQUAL(
       VariableFieldSizeCalculator::TOO_LARGE,
       m_calculator.CalculateFieldSize(
@@ -253,7 +253,7 @@ void VariableFieldSizeCalculatorTest::testSingleVariableSizedGroup() {
         3,
         &descriptor,
         &variable_field_size));
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(0), variable_field_size);
+  CPPUNIT_ASSERT_EQUAL(0u, variable_field_size);
   CPPUNIT_ASSERT_EQUAL(
       VariableFieldSizeCalculator::MISMATCHED_SIZE,
       m_calculator.CalculateFieldSize(
@@ -266,7 +266,7 @@ void VariableFieldSizeCalculatorTest::testSingleVariableSizedGroup() {
         5,
         &descriptor,
         &variable_field_size));
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(1), variable_field_size);
+  CPPUNIT_ASSERT_EQUAL(1u, variable_field_size);
   CPPUNIT_ASSERT_EQUAL(
       VariableFieldSizeCalculator::MISMATCHED_SIZE,
       m_calculator.CalculateFieldSize(
@@ -279,7 +279,7 @@ void VariableFieldSizeCalculatorTest::testSingleVariableSizedGroup() {
         7,
         &descriptor,
         &variable_field_size));
-  CPPUNIT_ASSERT_EQUAL(static_cast<unsigned int>(2), variable_field_size);
+  CPPUNIT_ASSERT_EQUAL(2u, variable_field_size);
 
   CPPUNIT_ASSERT_EQUAL(
       VariableFieldSizeCalculator::TOO_LARGE,
